Add logic to choose between interactive and non-interactive modes
Added logic in the main.js file to determine whether to run interactive mode or non-interactive mode based on the number of arguments provided on the command line. If the program is run with a file path, non-interactive mode is executed and if only one argument is provided, interactive mode is activated. This is done for ease of use of the program.